Replace jQuery .delegate() with .on()
jQuery has discouraged the use of .delegate()
in favor of .on() since the release of jQuery 1.7 in 2011.
In jQuery 3, .delegate() has been deprecated and
is up for removal. So jquery-migrate complains
about the usage of .delegate().

This change will replace delegate() with on(). That
also means jQuery versions < 1.7 will not be
supported, but supposedly it should be more
important to support the current version of jQuery
than to support the 6 year old version.
(At least in future releases?)